[feat/#6] Home 뷰모델 수정
