+++ app/src/main/java/com/sopt/dive/ui/feature/home/dummyData.kt
@@ -1,0 +1,65 @@
+package com.sopt.dive.ui.feature.home
+
+import com.sopt.dive.domain.model.ProfileInfo
+
+val dummyFriendList = listOf(
+    ProfileInfo(
+        profileImageUrl = "https://i.pinimg.com/736x/22/42/9b/22429b5137154dbefb84162d58d3f76b.jpg",
+        name = "이지현",
+        statusMessage = "안녕하세요",
+        music = "COLOR - NCT WISH"
+    ),
+    ProfileInfo(
+        name = "이지현",
+        isBirthday = true
+    ),
+    ProfileInfo(
+        profileImageUrl = "https://i.pinimg.com/736x/e6/17/fa/e617fa85c330cba8abbf56e11c542913.jpg",
+        name = "이지현",
+        music = "Drowning - WOODZ"
+    ),
+    ProfileInfo(
+        profileImageUrl = "https://i.pinimg.com/736x/22/42/9b/22429b5137154dbefb84162d58d3f76b.jpg",
+        name = "이지현",
+        statusMessage = "안녕하세요"
+    ),
+    ProfileInfo(
+        name = "이지현",
+        statusMessage = "생일이다아",
+        isBirthday = true
+    ),
+    ProfileInfo(
+        profileImageUrl = "https://i.pinimg.com/736x/e6/17/fa/e617fa85c330cba8abbf56e11c542913.jpg",
+        name = "이지현",
+        music = "FAMOUS - ALLDAY PROJECT"
+    ),
+    ProfileInfo(
+        profileImageUrl = "https://i.pinimg.com/736x/22/42/9b/22429b5137154dbefb84162d58d3f76b.jpg",
+        name = "이지현",
+        statusMessage = "안녕하세요",
+    ),
+    ProfileInfo(
+        name = "이지현",
+        isBirthday = true
+    ),
+    ProfileInfo(
+        profileImageUrl = "https://i.pinimg.com/736x/e6/17/fa/e617fa85c330cba8abbf56e11c542913.jpg",
+        name = "이지현",
+        isBirthday = true,
+        music = "Drowning - WOODZ"
+    ),
+    ProfileInfo(
+        profileImageUrl = "https://i.pinimg.com/736x/22/42/9b/22429b5137154dbefb84162d58d3f76b.jpg",
+        name = "이지현",
+        statusMessage = "안녕하세요",
+        music = "Blue Valentine - NMIXX"
+    ),
+    ProfileInfo(
+        name = "이지현",
+        isBirthday = true
+    ),
+    ProfileInfo(
+        profileImageUrl = "https://i.pinimg.com/736x/e6/17/fa/e617fa85c330cba8abbf56e11c542913.jpg",
+        name = "이지현"
+    )
+)
